Bug fix saving load/processor state Code cleanup

+++ src/FutureState.Flow/Core/ProcessorSingleResult.cs
@@ -17,42 +17,11 @@ namespace FutureState.Flow.Core
     public class ProcessorSingleResult<TEntityIn, TEntityOut> : IProcessor
         where TEntityOut : class, new()
     {
-        readonly Logger _logger;
-        private readonly Func<IEnumerable<TEntityIn>> _reader;
+        private readonly IEnumerable<ISpecification<IEnumerable<TEntityOut>>> _collectionRules;
+        private readonly Logger _logger;
         private readonly ObjectsMapper<TEntityIn, TEntityOut> _mapper;
+        private readonly Func<IEnumerable<TEntityIn>> _reader;
         private readonly IEnumerable<ISpecification<TEntityOut>> _rules;
-        private IEnumerable<ISpecification<IEnumerable<TEntityOut>>> _collectionRules;
-
-        /// <summary>
-        ///     Gets or sets the working folder to persist temporary files to.
-        /// </summary>
-        public string WorkingFolder { get; set; }
-
-        /// <summary>
-        ///     Gets the correlation id.
-        /// </summary>
-        public Guid CorrelationId { get; set; }
-        /// <summary>
-        ///     Gets the batch id.
-        /// </summary>
-        public int BatchId { get; set; }
-
-        /// <summary>
-        ///     Raised after processing.
-        /// </summary>
-
-        public event EventHandler<EventArgs> OnFinishedProcessing;
-
-        /// <summary>
-        ///     Called before processing an incoming dto to an outgoing dto.
-        /// </summary>
-        public Action<TEntityIn, TEntityOut> BeginProcessingItem { get; set; }
-
-        /// <summary>
-        ///     Get the well known name of the processor type.
-        /// </summary>
-
-        public string ProcessName { get; }
 
         /// <summary>
         ///     Creates a new instance.
@@ -76,12 +45,75 @@ namespace FutureState.Flow.Core
 
             _mapper = mapper ?? ObjectMapperManager.DefaultInstance.GetMapper<TEntityIn, TEntityOut>();
 
-            _rules = specProviderForEntity?.GetSpecifications().ToArray() ?? Enumerable.Empty<ISpecification<TEntityOut>>();
-            _collectionRules = specProviderForEntityCollection?.GetSpecifications().ToArray() ?? Enumerable.Empty<ISpecification<IEnumerable<TEntityOut>>>();
-
+            _rules = specProviderForEntity?.GetSpecifications().ToArray() ??
+                     Enumerable.Empty<ISpecification<TEntityOut>>();
+            _collectionRules = specProviderForEntityCollection?.GetSpecifications().ToArray() ??
+                               Enumerable.Empty<ISpecification<IEnumerable<TEntityOut>>>();
         }
 
-        protected IEnumerable<TEntityIn> Read() => _reader();
+        /// <summary>
+        ///     Gets or sets the working folder to persist temporary files to.
+        /// </summary>
+        public string WorkingFolder { get; set; }
+
+        /// <summary>
+        ///     Gets the correlation id.
+        /// </summary>
+        public Guid CorrelationId { get; set; }
+
+        /// <summary>
+        ///     Gets the batch id.
+        /// </summary>
+        public int BatchId { get; set; }
+
+        /// <summary>
+        ///     Called before processing an incoming dto to an outgoing dto.
+        /// </summary>
+        public Action<TEntityIn, TEntityOut> BeginProcessingItem { get; set; }
+
+        /// <summary>
+        ///     Gets the errors encountered processing from the incoming data source.
+        /// </summary>
+        public List<ProcessError<TEntityIn>> ProcessedErrors { get; private set; }
+
+        /// <summary>
+        ///     Gets/sets the processor handler.
+        /// </summary>
+        public ProcessorHandler<TEntityIn> ProcessorHandler { get; private set; }
+
+        /// <summary>
+        ///     Gets the valid items created after processing.
+        /// </summary>
+        public List<TEntityOut> ProcessedItems { get; private set; }
+
+        /// <summary>
+        ///     Get the well known name of the processor type.
+        /// </summary>
+
+        public string ProcessName { get; }
+
+        /// <summary>
+        ///     Processes a snapshot from an incoming source.
+        /// </summary>
+        public ProcessResult Process()
+        {
+            if (!Directory.Exists(WorkingFolder))
+                Directory.CreateDirectory(WorkingFolder);
+
+            var handler = CreateProcessHandler();
+
+            return Process(handler);
+        }
+
+        /// <summary>
+        ///     Raised after processing.
+        /// </summary>
+        public event EventHandler<EventArgs> OnFinishedProcessing;
+
+        protected IEnumerable<TEntityIn> Read()
+        {
+            return _reader();
+        }
 
         /// <summary>
         ///     Creates a new process hanlder.
@@ -91,12 +123,12 @@ namespace FutureState.Flow.Core
         {
             var processedValidItems = new List<TEntityOut>();
 
-            var pHandler = new ProcessorHandler<TEntityIn>(ProcessName,CorrelationId, BatchId)
+            var pHandler = new ProcessorHandler<TEntityIn>(ProcessName, CorrelationId, BatchId)
             {
                 EntitiesReader = Read(),
                 Logger = _logger,
                 OnError = OnError,
-                ProcessItem = (dtoIn) =>
+                ProcessItem = dtoIn =>
                 {
                     // create output entity
                     var dtoOut = new TEntityOut();
@@ -105,9 +137,10 @@ namespace FutureState.Flow.Core
                     dtoOut = _mapper.Map(dtoIn, dtoOut);
 
                     // prepare entity
-                    BeginProcessingItem?.Invoke(dtoIn, dtoOut); ;
+                    BeginProcessingItem?.Invoke(dtoIn, dtoOut);
+                    ;
 
-                    ErrorEvent errorEvent = OnItemProcessing(dtoIn, dtoOut);
+                    var errorEvent = OnItemProcessing(dtoIn, dtoOut);
 
                     // validate against business rules
                     if (errorEvent == null)
@@ -117,7 +150,7 @@ namespace FutureState.Flow.Core
                         if (e.Any())
                         {
                             var first = e.First();
-                            errorEvent = new ErrorEvent() { Message = first.Message, Type = first.Type };
+                            errorEvent = new ErrorEvent {Message = first.Message, Type = first.Type};
                         }
                     }
 
@@ -131,31 +164,15 @@ namespace FutureState.Flow.Core
                     var errors = _collectionRules.ToErrors(processedValidItems);
                     var enumerable = errors as Error[] ?? errors.ToArray();
                     if (enumerable.Any())
-                        throw new RuleException("Unable to commit items due to one or more rules. Please see the inner exception for more details.", enumerable);
+                        throw new RuleException(
+                            "Unable to commit items due to one or more rules. Please see the inner exception for more details.",
+                            enumerable);
 
                     Commit(processedValidItems);
                 }
             };
 
             return pHandler;
-        }
-
-        /// <summary>
-        ///     Gets the errors encountered processing from the incoming data source.
-        /// </summary>
-        public List<ProcessError<TEntityIn>> ProcessedErrors { get; private set; }
-
-        /// <summary>
-        ///     Processes a snapshot from an incoming source.
-        /// </summary>
-        public ProcessResult Process()
-        {
-            if (!Directory.Exists(WorkingFolder))
-                Directory.CreateDirectory(WorkingFolder);
-
-            ProcessorHandler<TEntityIn> handler = CreateProcessHandler();
-
-            return Process(handler);
         }
 
         /// <summary>
@@ -167,24 +184,19 @@ namespace FutureState.Flow.Core
         {
             Guard.ArgumentNotNull(processHandler, nameof(processHandler));
 
-            this.ProcessorHandler = processHandler;
+            ProcessorHandler = processHandler;
 
-            this.ProcessedItems = new List<TEntityOut>(); // clear
+            ProcessedItems = new List<TEntityOut>(); // clear
 
-            ProcessResult result = processHandler.Process();
+            var result = processHandler.Process();
 
-            this.ProcessedErrors = processHandler.Errors;
+            ProcessedErrors = processHandler.Errors;
 
             // raise event
             OnFinishedProcessing?.Invoke(this, EventArgs.Empty);
 
             return result;
         }
-
-        /// <summary>
-        ///     Gets/sets the processor handler.
-        /// </summary>
-        public ProcessorHandler<TEntityIn> ProcessorHandler { get; private set; }
 
         /// <summary>
         ///     Processes a single item and returns and error event.
@@ -204,19 +216,14 @@ namespace FutureState.Flow.Core
             var i = 1;
             var batchAsList = batch.ToList();
 
-            var fileName = $@"{WorkingFolder}\{ProcessName}-OnFinishedProcessing-{CorrelationId}-{BatchId}.json";
+            var fileName = $@"{WorkingFolder}\{ProcessName}-{CorrelationId}-{BatchId}.json";
             while (File.Exists(fileName))
                 fileName = $@"{WorkingFolder}\{ProcessName}-{CorrelationId}-{BatchId}-{i++}.json";
 
             SaveSnapShot(fileName, batchAsList);
 
-            this.ProcessedItems = batchAsList;
+            ProcessedItems = batchAsList;
         }
-
-        /// <summary>
-        ///     Gets the valid items created after processing.
-        /// </summary>
-        public List<TEntityOut> ProcessedItems { get; private set; }
 
         /// <summary>
         ///     Raised whenever an error occured processign the results.
@@ -225,7 +232,6 @@ namespace FutureState.Flow.Core
         /// <param name="error">The error exception raised.</param>
         public virtual void OnError(TEntityIn entityIn, Exception error)
         {
-
         }
 
         /// <summary>
@@ -233,11 +239,7 @@ namespace FutureState.Flow.Core
         /// </summary>
         private void SaveSnapShot<T>(string filePath, List<T> data)
         {
-            if (File.Exists(filePath))
-            {
-                // todo: log result
-                File.Delete(filePath);
-            }
+            if (File.Exists(filePath)) File.Delete(filePath);
 
             var log = new ProcessSnapshot<T>
             {
